Move worker process into separate file
